+++ src/Concordium/Client/Types/Transaction.hs
@@ -108,7 +108,7 @@ bakerConfigurePayloadSize hasCapital hasRestake hasPoolOpen hasKeys mMetadata ha
   + (if hasRestake then 1 else 0)
   + (if hasPoolOpen then 1 else 0)
   + (if hasKeys then fromIntegral bakerKeysWithProofsSize else 0)
-  + maybe 0 fromIntegral mMetadata
+  + maybe 0 ((+2) . fromIntegral) mMetadata
   + (if hasTCom then 4 else 0)
   + (if hasBCom then 4 else 0)
   + (if hasFCom then 4 else 0)