+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 - Fix calculation of payload size of a configure baker transaction in
-  the function `bakerConfigurePayloadSize` in Transactions.hs used by
+  the function `bakerConfigurePayloadSize` in Transaction.hs used by
   the wallet-proxy.
 
 ## 4.0.2